small fixes and tests added

+++ src/models/message.rs
@@ -3,7 +3,7 @@ use crate::models::signal::Signal;
 #[derive(Default, Clone)]
 // BO_ <ID> <MESSAGE_NAME> : <BYTES_LENGHT> <SENDER_NODE>
 pub struct Message {
-    pub id: u32,
+    pub id: u64,
     pub id_hex: String,
     pub name: String,
     pub byte_length: usize,
